<Feature>(Logic): Implementation of  a high security possible password creator
